test: add redis `Purge` api test
