+++ internal/database/online/redis/import.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	"github.com/oom-ai/oomstore/internal/database/online"
-	"github.com/oom-ai/oomstore/pkg/oomstore/types"
 )
 
 func (db *DB) Import(ctx context.Context, opt online.ImportOpt) error {
@@ -60,31 +59,6 @@ func (db *DB) Import(ctx context.Context, opt online.ImportOpt) error {
 	if seq%PipelineBatchSize != 0 {
 		if _, err := pipe.Exec(ctx); err != nil {
 			return err
-		}
-	}
-	return nil
-}
-
-func (db *DB) Purge(ctx context.Context, revision *types.Revision) error {
-	prefix, err := SerializeByValue(revision.ID)
-	if err != nil {
-		return nil
-	}
-	pattern := prefix + ":*"
-
-	var cursor uint64
-	for {
-		keys, cursor, err := db.Scan(ctx, cursor, pattern, PipelineBatchSize).Result()
-		if err != nil {
-			return err
-		}
-
-		if _, err = db.Del(ctx, keys...).Result(); err != nil {
-			return err
-		}
-
-		if cursor == 0 {
-			break
 		}
 	}
 	return nil
